middleware rework
